@@ -4,19 +4,19 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'de.felixschulze.gradle.hockeyapp'
 apply from: '../quality/quality.gradle'
-apply plugin: 'io.objectbox'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'io.objectbox'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
     defaultConfig {
         applicationId "de.bornholdtlee.baseproject"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode System.getenv("CI_PIPELINE_ID") as Integer ?: 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         buildConfigField 'String', 'AUTH_USER', '"BLG"'
         buildConfigField 'String', 'AUTH_PW', '"SuperStrongPW"'
@@ -131,15 +131,15 @@ dependencies {
 
     /////UI/////
 
-    implementation 'com.google.android.gms:play-services-maps:15.0.1'
+    implementation 'com.google.android.gms:play-services-maps:16.0.0'
     implementation 'com.google.maps.android:android-maps-utils:0.5'
 
-    implementation "com.android.support:appcompat-v7:$supportVersion"
-    implementation "com.android.support:support-v4:$supportVersion"
-    implementation "com.android.support:design:$supportVersion"
-    implementation "com.android.support:cardview-v7:$supportVersion"
-    implementation "com.android.support:recyclerview-v7:$supportVersion"
-    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.1.0-alpha02'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0-alpha01'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha3'
 
     implementation 'com.aurelhubert:ahbottomnavigation:2.1.0'
 
@@ -148,11 +148,11 @@ dependencies {
     /////SUPPORT/////
 
     //Permissions
-    implementation 'com.karumi:dexter:4.2.0'
+    implementation 'com.karumi:dexter:5.0.0'
     //Date
-    implementation 'joda-time:joda-time:2.9.9'
+    implementation 'joda-time:joda-time:2.10.1'
 
-    implementation 'net.hockeyapp.android:HockeySDK:5.0.4'
+    implementation 'net.hockeyapp.android:HockeySDK:5.1.1'
     implementation "com.facebook.stetho:stetho:$stethoVersion"
     implementation "com.facebook.stetho:stetho-okhttp3:$stethoVersion"
 
@@ -172,8 +172,8 @@ dependencies {
     annotationProcessor "com.jakewharton:butterknife-compiler:$butterknifeVersion"
     kapt "com.jakewharton:butterknife-compiler:$butterknifeVersion"
 
-    implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    implementation 'io.reactivex.rxjava2:rxjava:2.1.6'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.3'
 
     /////DATABASE/////
 
@@ -192,9 +192,8 @@ dependencies {
     /////TESTING/////
 
     androidTestImplementation 'com.android.support.test:testing-support-lib:0.1'
-    testImplementation 'org.testng:testng:6.9.6'
     testImplementation 'junit:junit:4.12'
-    testImplementation "org.mockito:mockito-inline:2.8.9"
+    testImplementation "org.mockito:mockito-inline:2.21.0"
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
     testImplementation "org.robolectric:shadows-multidex:$robolectricVersion"
     testImplementation "org.powermock:powermock-module-junit4:$powermockVersion"
