@@ -2,13 +2,13 @@
 
 buildscript {
     ext{
-        kotlin_version = '1.2.41'
+        kotlin_version = '1.3.20'
 
         daggerVersion = '2.15'
-        butterknifeVersion = '8.8.1'
+        butterknifeVersion = '10.0.0'
 
         supportVersion = '27.1.1'
-        objectboxVersion = '1.5.0'
+        objectboxVersion = '2.3.1'
         retrofitVersion = '2.4.0'
 
         powermockVersion = '1.7.3'
@@ -18,14 +18,15 @@ buildscript {
     }
 
     repositories {
+        google()
         jcenter()
         maven {
             url "https://plugins.gradle.org/m2/"
         }
-        google()
     }
+
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
+        classpath 'com.android.tools.build:gradle:3.3.0'
         classpath 'de.felixschulze.gradle:gradle-hockeyapp-plugin:3.6'
         classpath "io.objectbox:objectbox-gradle-plugin:$objectboxVersion"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
