@@ -3,8 +3,8 @@ package de.bornholdtlee.baseproject.base
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.Toolbar
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import android.view.*
 import android.view.View.GONE
 import android.view.View.VISIBLE
@@ -15,7 +15,10 @@ import de.bornholdtlee.baseproject.base.navigation.NavigationBaseActivity
 import de.bornholdtlee.baseproject.enums.AnimationType
 import de.bornholdtlee.baseproject.injection.IInjection
 import de.bornholdtlee.baseproject.injection.components.AppComponent
+import de.bornholdtlee.baseproject.model.Test.DelegateShit
+import de.bornholdtlee.baseproject.model.Test.TesteMich
 import de.bornholdtlee.baseproject.ui.main.MainFragment
+import de.bornholdtlee.baseproject.utils.Logger
 import de.bornholdtlee.baseproject.utils.UiUtils
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper
 import javax.inject.Inject
@@ -69,6 +72,13 @@ abstract class BaseActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        var delegate : DelegateShit = DelegateShit()
+        var test: TesteMich = TesteMich(delegate)
+
+        Logger.error(test.method1())
+        Logger.error(test.method2())
+
 
         if (this is IInjection) {
             (this as IInjection).inject(appComponent)
@@ -160,7 +170,7 @@ abstract class BaseActivity : AppCompatActivity() {
         val fragmentName: String = fragment.javaClass.name
 
         if (clearTop) { //maybe not functional TODO
-            val fragmentPopped: Boolean = fragmentManager.popBackStackImmediate(fragmentName, 0)
+            val fragmentPopped: Boolean = supportFragmentManager.popBackStackImmediate(fragmentName, 0)
             if (fragmentPopped) {
                 return
             }
