@@ -2,7 +2,7 @@ package de.bornholdtlee.baseproject.viewbuilder
 
 import android.content.Context
 import android.content.DialogInterface
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import de.bornholdtlee.baseproject.R
 
 class DialogBuilder {
