@@ -1,10 +1,10 @@
 package de.bornholdtlee.baseproject.base.recyclerview
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
 
 abstract class BaseSimpleRecyclerAdapter<M, H : BaseViewHolder<M>>(context: Context, private val presenter: IAdapterPresenter<M>) : RecyclerView.Adapter<H>() {
     private val inflater: LayoutInflater = LayoutInflater.from(context)
