@@ -1,9 +1,9 @@
 package de.bornholdtlee.baseproject.ui.creation
 
 import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.View
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import butterknife.BindView
 import de.bornholdtlee.baseproject.R
 import de.bornholdtlee.baseproject.model.Lesson
