@@ -3,8 +3,6 @@ package de.bornholdtlee.baseproject.utils
 import android.app.Activity
 import android.content.Context
 import android.content.res.Resources
-import android.support.v4.app.FragmentTransaction
-import android.support.v4.content.ContextCompat
 import android.util.DisplayMetrics
 import android.view.View
 import android.view.animation.Animation
@@ -13,6 +11,8 @@ import android.view.animation.RotateAnimation
 import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.FragmentTransaction
 
 import de.bornholdtlee.baseproject.R
 import de.bornholdtlee.baseproject.base.BaseActivity
